pyproject.toml
[project]
description = "Protostar Cairo bindings for Python"
name = "protostar_cairo_py"
readme = "README.md"
version = "0.0.1"

[tool.poetry]
authors = []
description = ""
name = ""
readme = ""
version = "0.0.0"

[tool.poe.tasks.build]
shell = "rustup override set nightly && maturin build --out ./build"

[tool.poe.tasks.sync]
shell = "git submodule update --remote --merge"

[tool.poe.tasks.deploy]
shell = "python scripts/deploy.py"

[tool.poetry.dependencies]
python = ">=3.9.14, <3.10"

[tool.poetry.group.dev.dependencies]
gitpython = "^3.1.30"
maturin = "^0.14.13"
poethepoet = "^0.18.1"
tomli = "^2.0.1"
tomli-w = "^1.0.0"

[build-system]
build-backend = "maturin"
requires = ["maturin>=0.14,<0.15"]

[tool.maturin]
exclude = ["build", "scripts"]
manifest-path = "rust/crates/cairo-lang-python-bindings/Cargo.toml"
profile = "release"
strip = true
